Code to create schedule, file makeSchedule.js
This is my code to generate the schedule. It takes it the members array. If the number of roles is n, it takes the first n members and permutes them to find a combination that has no conflicts (as defined by each members array of exemptions). If a schedule is found with no conflicts, it is organized into the correct schedule format and returned. If no "perfect" schedule is found, the best fit is returned.

I built this with what I understand to be the input and output requirements, but I don't have a working database so I can't test it.